Change "kernel" to "operating sytstem" in markdown readme
This is how you know I'm a Linux guy.
I did not realise this was the whole OS with basic user-space programs
like 'cat' and everything. I thought this was just the OS kernel, like
"Linux" is.

Also, if you take the time to read this commit message, you know you
shouln't trust me to give you the source for your OS.

But hey, *hugs* :-)